+++ CHANGELOG.md
@@ -8,10 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 - Fix GitHub workflows continuous integration.
-
 - Make private data structures that are private and subclass private
   respectfully: `Dealloc.ALLOCATIONS` and
   `PersistableContainer.PERSISTABLE_TRANSIENT_ATTRIBUTE`.
+- Fix to `FactoryStash` for persisting `None` to file system instead of using
+  the factory stash to create the item.
 
 
 ## [1.5.3] - 2021-08-16